Attempt to fix introduced alert

+++ ncdump/ncvalidator.c
@@ -578,7 +578,7 @@ hdr_len_NC_var(const NC_var *varp,
      */
     sz = hdr_len_NC_name(varp->name_len, sizeof_t);     /* name */
     sz += sizeof_t;                                     /* nelems */
-    sz += sizeof_t * varp->ndims;                       /* [dimid ...] */
+    sz += ((long long)sizeof_t) * varp->ndims;          /* [dimid ...] */
     sz += hdr_len_NC_attrarray(&varp->attrs, sizeof_t); /* vatt_list */
     sz += 4;                                            /* nc_type */
     sz += sizeof_t;                                     /* vsize */
